GLOBAL: Introduce loadtextureimage unification
