+++ source/ctr/gl/gl_draw.c
@@ -225,7 +225,7 @@ qpic_t	*Draw_CachePic (char *path)
 // load the pic from disk
 //
 
-	int index = loadtextureimage (str, 0, 0, qfalse, qfalse);
+	int index = loadtextureimage (str, 0, 0, qfalse, 0, qfalse qfalse);
 	if(index)
 	{
 		pic->pic.width  = gltextures[index].original_width;
@@ -383,7 +383,7 @@ void Draw_Init (void)
 	Cmd_AddCommand ("gl_texturemode", &Draw_TextureMode_f);
 
 	// now turn them into textures
-	char_texture = loadtextureimage ("gfx/charset", 0, 0, false, false);
+	char_texture = loadtextureimage ("gfx/charset", 0, 0, false, 0, false, false);
 	if (char_texture == 0)// did not find a matching TGA...
 		Sys_Error ("Could not load charset, make sure you have every folder and file installed properly\nDouble check that all of your files are in their correct places\nAnd that you have installed the game properly.\n");
 
@@ -1780,557 +1780,3 @@ void GL_SelectTexture (GLenum target)
 	currenttexture = cnttextures[target-TEXTURE0_SGIS];
 	oldtarget = target;
 }
-
-
-
-//Diabolickal TGA Begin
-int		image_width;
-int		image_height;
-#define	IMAGE_MAX_DIMENSIONS	512
-
-/*
-=================================================================
-  PCX Loading
-=================================================================
-*/
-
-typedef struct
-{
-    char	manufacturer;
-    char	version;
-    char	encoding;
-    char	bits_per_pixel;
-    unsigned short	xmin,ymin,xmax,ymax;
-    unsigned short	hres,vres;
-    unsigned char	palette[48];
-    char	reserved;
-    char	color_planes;
-    unsigned short	bytes_per_line;
-    unsigned short	palette_type;
-    char	filler[58];
-    unsigned 	data;			// unbounded
-} pcx_t;
-
-/*
-============
-LoadPCX
-============
-*/
-byte* LoadPCX (FILE *f, int matchwidth, int matchheight)
-{
-	pcx_t	*pcx, pcxbuf;
-	byte	palette[768];
-	byte	*pix, *image_rgba;
-	int		x, y;
-	int		dataByte, runLength;
-	int		count;
-
-//
-// parse the PCX file
-//
-	fread (&pcxbuf, 1, sizeof(pcxbuf), f);
-	pcx = &pcxbuf;
-
-	if (pcx->manufacturer != 0x0a
-		|| pcx->version != 5
-		|| pcx->encoding != 1
-		|| pcx->bits_per_pixel != 8
-		|| pcx->xmax >= 514
-		|| pcx->ymax >= 514)
-	{
-		Con_Printf ("Bad pcx file\n");
-		return NULL;
-	}
-	if (matchwidth && (pcx->xmax+1) != matchwidth)
-		return NULL;
-	if (matchheight && (pcx->ymax+1) != matchheight)
-		return NULL;
-	// seek to palette
-	fseek (f, -768, SEEK_END);
-	fread (palette, 1, 768, f);
-	fseek (f, sizeof(pcxbuf) - 4, SEEK_SET);
-	count = (pcx->xmax+1) * (pcx->ymax+1);
-	image_rgba = (byte*)malloc( count * 4);
-
-	for (y=0 ; y<=pcx->ymax ; y++)
-	{
-		pix = image_rgba + 4*y*(pcx->xmax+1);
-		for (x=0 ; x<=pcx->xmax ; ) // muff - fixed - was referencing ymax
-		{
-			dataByte = fgetc(f);
-			if((dataByte & 0xC0) == 0xC0)
-			{
-				runLength = dataByte & 0x3F;
-				dataByte = fgetc(f);
-			}
-			else
-				runLength = 1;
-
-			while(runLength-- > 0)
-			{
-				pix[0] = palette[dataByte*3];
-				pix[1] = palette[dataByte*3+1];
-				pix[2] = palette[dataByte*3+2];
-				pix[3] = 255;
-				pix += 4;
-				x++;
-			}
-		}
-	}
-	image_width = pcx->xmax+1;
-	image_height = pcx->ymax+1;
-
-	fclose(f);
-	return image_rgba;
-}
-
-
-/*
-=========================================================
-
-			Targa
-
-=========================================================
-*/
-
-#define TGA_MAXCOLORS 16384
-
-/* Definitions for image types. */
-#define TGA_Null	0	/* no image data */
-#define TGA_Map		1	/* Uncompressed, color-mapped images. */
-#define TGA_RGB		2	/* Uncompressed, RGB images. */
-#define TGA_Mono	3	/* Uncompressed, black and white images. */
-#define TGA_RLEMap	9	/* Runlength encoded color-mapped images. */
-#define TGA_RLERGB	10	/* Runlength encoded RGB images. */
-#define TGA_RLEMono	11	/* Compressed, black and white images. */
-#define TGA_CompMap	32	/* Compressed color-mapped data, using Huffman, Delta, and runlength encoding. */
-#define TGA_CompMap4	33	/* Compressed color-mapped data, using Huffman, Delta, and runlength encoding. 4-pass quadtree-type process. */
-
-/* Definitions for interleave flag. */
-#define TGA_IL_None	0	/* non-interleaved. */
-#define TGA_IL_Two	1	/* two-way (even/odd) interleaving */
-#define TGA_IL_Four	2	/* four way interleaving */
-#define TGA_IL_Reserved	3	/* reserved */
-
-/* Definitions for origin flag */
-#define TGA_O_UPPER	0	/* Origin in lower left-hand corner. */
-#define TGA_O_LOWER	1	/* Origin in upper left-hand corner. */
-
-typedef struct _TargaHeader
-{
-	unsigned char 	id_length, colormap_type, image_type;
-	unsigned short	colormap_index, colormap_length;
-	unsigned char	colormap_size;
-	unsigned short	x_origin, y_origin, width, height;
-	unsigned char	pixel_size, attributes;
-} TargaHeader;
-
-int fgetLittleShort (FILE *f)
-{
-	byte	b1, b2;
-
-	b1 = fgetc(f);
-	b2 = fgetc(f);
-
-	return (short)(b1 + b2*256);
-}
-
-int fgetLittleLong (FILE *f)
-{
-	byte	b1, b2, b3, b4;
-
-	b1 = fgetc(f);
-	b2 = fgetc(f);
-	b3 = fgetc(f);
-	b4 = fgetc(f);
-
-	return b1 + (b2<<8) + (b3<<16) + (b4<<24);
-}
-
-/*
-=============
-LoadTGA
-=============
-*/
-byte *LoadTGA (FILE *fin, int matchwidth, int matchheight)
-{
-	int		w, h, x, y, realrow, truerow, baserow, i, temp1, temp2, pixel_size, map_idx;
-	int		RLE_count, RLE_flag, size, interleave, origin;
-	qboolean	mapped, rlencoded;
-	byte		*data, *dst, r, g, b, a, j, k, l, *ColorMap;
-	TargaHeader	header;
-
-	header.id_length = fgetc (fin);
-	header.colormap_type = fgetc (fin);
-	header.image_type = fgetc (fin);
-	header.colormap_index = fgetLittleShort (fin);
-	header.colormap_length = fgetLittleShort (fin);
-	header.colormap_size = fgetc (fin);
-	header.x_origin = fgetLittleShort (fin);
-	header.y_origin = fgetLittleShort (fin);
-	header.width = fgetLittleShort (fin);
-	header.height = fgetLittleShort (fin);
-	header.pixel_size = fgetc (fin);
-	header.attributes = fgetc (fin);
-
-	if (header.width > IMAGE_MAX_DIMENSIONS || header.height > IMAGE_MAX_DIMENSIONS)
-	{
-		Con_DPrintf ("TGA image %s exceeds maximum supported dimensions\n", fin);
-		fclose (fin);
-		return NULL;
-	}
-
-	if ((matchwidth && header.width != matchwidth) || (matchheight && header.height != matchheight))
-	{
-		fclose (fin);
-		return NULL;
-	}
-
-	if (header.id_length != 0)
-		fseek (fin, header.id_length, SEEK_CUR);
-
-	/* validate TGA type */
-	switch (header.image_type)
-	{
-	case TGA_Map:
-	case TGA_RGB:
-	case TGA_Mono:
-	case TGA_RLEMap:
-	case TGA_RLERGB:
-	case TGA_RLEMono:
-		break;
-
-	default:
-		Con_DPrintf ("Unsupported TGA image %s: Only type 1 (map), 2 (RGB), 3 (mono), 9 (RLEmap), 10 (RLERGB), 11 (RLEmono) TGA images supported\n");
-		fclose (fin);
-		return NULL;
-	}
-
-	/* validate color depth */
-	switch (header.pixel_size)
-	{
-	case 8:
-	case 15:
-	case 16:
-	case 24:
-	case 32:
-		break;
-
-	default:
-		Con_DPrintf ("Unsupported TGA image %s: Only 8, 15, 16, 24 or 32 bit images (with colormaps) supported\n");
-		fclose (fin);
-		return NULL;
-	}
-
-	r = g = b = a = l = 0;
-
-	/* if required, read the color map information. */
-	ColorMap = NULL;
-	mapped = (header.image_type == TGA_Map || header.image_type == TGA_RLEMap) && header.colormap_type == 1;
-	if (mapped)
-	{
-		/* validate colormap size */
-		switch (header.colormap_size)
-		{
-		case 8:
-		case 15:
-		case 16:
-		case 32:
-		case 24:
-			break;
-
-		default:
-			Con_DPrintf ("Unsupported TGA image %s: Only 8, 15, 16, 24 or 32 bit colormaps supported\n");
-			fclose (fin);
-			return NULL;
-		}
-
-		temp1 = header.colormap_index;
-		temp2 = header.colormap_length;
-		if ((temp1 + temp2 + 1) >= TGA_MAXCOLORS)
-		{
-			fclose (fin);
-			return NULL;
-		}
-		ColorMap = (byte*)(malloc (TGA_MAXCOLORS * 4));
-		map_idx = 0;
-		for (i = temp1 ; i < temp1 + temp2 ; ++i, map_idx += 4)
-		{
-			/* read appropriate number of bytes, break into rgb & put in map. */
-			switch (header.colormap_size)
-			{
-			case 8:	/* grey scale, read and triplicate. */
-				r = g = b = getc (fin);
-				a = 255;
-				break;
-
-			case 15:	/* 5 bits each of red green and blue. */
-						/* watch byte order. */
-				j = getc (fin);
-				k = getc (fin);
-				l = ((unsigned int)k << 8) + j;
-				r = (byte)(((k & 0x7C) >> 2) << 3);
-				g = (byte)((((k & 0x03) << 3) + ((j & 0xE0) >> 5)) << 3);
-				b = (byte)((j & 0x1F) << 3);
-				a = 255;
-				break;
-
-			case 16:	/* 5 bits each of red green and blue, 1 alpha bit. */
-						/* watch byte order. */
-				j = getc (fin);
-				k = getc (fin);
-				l = ((unsigned int)k << 8) + j;
-				r = (byte)(((k & 0x7C) >> 2) << 3);
-				g = (byte)((((k & 0x03) << 3) + ((j & 0xE0) >> 5)) << 3);
-				b = (byte)((j & 0x1F) << 3);
-				a = (k & 0x80) ? 255 : 0;
-				break;
-
-			case 24:	/* 8 bits each of blue, green and red. */
-				b = getc (fin);
-				g = getc (fin);
-				r = getc (fin);
-				a = 255;
-				l = 0;
-				break;
-
-			case 32:	/* 8 bits each of blue, green, red and alpha. */
-				b = getc (fin);
-				g = getc (fin);
-				r = getc (fin);
-				a = getc (fin);
-				l = 0;
-				break;
-			}
-			ColorMap[map_idx+0] = r;
-			ColorMap[map_idx+1] = g;
-			ColorMap[map_idx+2] = b;
-			ColorMap[map_idx+3] = a;
-		}
-	}
-
-	/* check run-length encoding. */
-	rlencoded = (header.image_type == TGA_RLEMap || header.image_type == TGA_RLERGB || header.image_type == TGA_RLEMono);
-	RLE_count = RLE_flag = 0;
-
-	image_width = w = header.width;
-	image_height = h = header.height;
-
-	size = w * h * 4;
-	data = (byte*)(malloc (size));
-
-	/* read the Targa file body and convert to portable format. */
-	pixel_size = header.pixel_size;
-	origin = (header.attributes & 0x20) >> 5;
-	interleave = (header.attributes & 0xC0) >> 6;
-	truerow = baserow = 0;
-	for (y=0 ; y<h ; y++)
-	{
-		realrow = truerow;
-		if (origin == TGA_O_UPPER)
-			realrow = h - realrow - 1;
-
-		dst = data + realrow * w * 4;
-
-		for (x=0 ; x<w ; x++)
-		{
-			/* check if run length encoded. */
-			if (rlencoded)
-			{
-				if (!RLE_count)
-				{
-					/* have to restart run. */
-					i = getc (fin);
-					RLE_flag = (i & 0x80);
-					if (!RLE_flag)	// stream of unencoded pixels
-						RLE_count = i + 1;
-					else		// single pixel replicated
-						RLE_count = i - 127;
-					/* decrement count & get pixel. */
-					--RLE_count;
-				}
-				else
-				{
-					/* have already read count & (at least) first pixel. */
-					--RLE_count;
-					if (RLE_flag)
-						/* replicated pixels. */
-						goto PixEncode;
-				}
-			}
-
-			/* read appropriate number of bytes, break into RGB. */
-			switch (pixel_size)
-			{
-			case 8:	/* grey scale, read and triplicate. */
-				r = g = b = l = getc (fin);
-				a = 255;
-				break;
-
-			case 15:	/* 5 bits each of red green and blue. */
-						/* watch byte order. */
-				j = getc (fin);
-				k = getc (fin);
-				l = ((unsigned int)k << 8) + j;
-				r = (byte)(((k & 0x7C) >> 2) << 3);
-				g = (byte)((((k & 0x03) << 3) + ((j & 0xE0) >> 5)) << 3);
-				b = (byte)((j & 0x1F) << 3);
-				a = 255;
-				break;
-
-			case 16:	/* 5 bits each of red green and blue, 1 alpha bit. */
-						/* watch byte order. */
-				j = getc (fin);
-				k = getc (fin);
-				l = ((unsigned int)k << 8) + j;
-				r = (byte)(((k & 0x7C) >> 2) << 3);
-				g = (byte)((((k & 0x03) << 3) + ((j & 0xE0) >> 5)) << 3);
-				b = (byte)((j & 0x1F) << 3);
-				a = (k & 0x80) ? 255 : 0;
-				break;
-
-			case 24:	/* 8 bits each of blue, green and red. */
-				b = getc (fin);
-				g = getc (fin);
-				r = getc (fin);
-				a = 255;
-				l = 0;
-				break;
-
-			case 32:	/* 8 bits each of blue, green, red and alpha. */
-				b = getc (fin);
-				g = getc (fin);
-				r = getc (fin);
-				a = getc (fin);
-				l = 0;
-				break;
-
-			default:
-				Con_DPrintf ("Malformed TGA image: Illegal pixel_size '%d'\n", pixel_size);
-				fclose (fin);
-				free (data);
-				if (mapped)
-					free (ColorMap);
-				return NULL;
-			}
-
-PixEncode:
-			if (mapped)
-			{
-				map_idx = l * 4;
-				*dst++ = ColorMap[map_idx+0];
-				*dst++ = ColorMap[map_idx+1];
-				*dst++ = ColorMap[map_idx+2];
-				*dst++ = ColorMap[map_idx+3];
-			}
-			else
-			{
-				*dst++ = r;
-				*dst++ = g;
-				*dst++ = b;
-				*dst++ = a;
-			}
-		}
-
-		if (interleave == TGA_IL_Four)
-			truerow += 4;
-		else if (interleave == TGA_IL_Two)
-			truerow += 2;
-		else
-			truerow++;
-		if (truerow >= h)
-			truerow = ++baserow;
-	}
-
-	if (mapped)
-		free (ColorMap);
-
-	fclose (fin);
-
-	return data;
-}
-
-/*small function to read files with stb_image - single-file image loader library.
-** downloaded from: https://raw.githubusercontent.com/nothings/stb/master/stb_image.h
-** only use jpeg+png formats, because tbh there's not much need for the others.
-** */
-#define STB_IMAGE_IMPLEMENTATION
-#define STBI_ONLY_JPEG
-#define STBI_ONLY_PNG
-#include "../../stb_image.h"
-byte* LoadSTBI(FILE *f, int width, int height)
-{
-	int bpp;
-	int inwidth, inheight;
-	byte* image = stbi_load_from_file(f, &inwidth, &inheight, &bpp, 4);
-	// wtf?
-	image_width = inwidth;
-	image_height = inheight;
-	fclose(f);
-	return image;
-}
-
-byte* loadimagepixels (char* filename, qboolean complain, int matchwidth, int matchheight)
-
-{
-	FILE	*f;
-	char	basename[128], name[132];
-	byte	*c;
-
-	if (complain == qfalse)
-		COM_StripExtension(filename, basename); // strip the extension to allow TGA
-	else
-		strcpy(basename, filename);
-
-	c = (byte*)basename;
-	while (*c)
-	{
-		if (*c == '*')
-			*c = '+';
-		c++;
-	}
-
-	//Try PCX
-	sprintf (name, "%s.pcx", basename);
-	COM_FOpenFile (name, &f);
-	if (f)
-		return LoadPCX (f, matchwidth, matchheight);
-	//Try TGA
-	sprintf (name, "%s.tga", basename);
-	COM_FOpenFile (name, &f);
-	if (f)
-		return LoadTGA (f, matchwidth, matchheight);
-	//Try PNG
-	sprintf (name, "%s.png", basename);
-	COM_FOpenFile (name, &f);
-	if (f)
-		return LoadSTBI (f, matchwidth, matchheight);
-	//Try JPEG
-	sprintf (name, "%s.jpeg", basename);
-	COM_FOpenFile (name, &f);
-	if (f)
-		return LoadSTBI (f, matchwidth, matchheight);
-	sprintf (name, "%s.jpg", basename);
-	COM_FOpenFile (name, &f);
-	if (f)
-		return LoadSTBI (f, matchwidth, matchheight);
-	
-	//if (complain)
-	//	Con_Printf ("Couldn't load %s.tga or %s.pcx \n", filename);
-	
-	return NULL;
-}
-
-int loadtextureimage (char* filename, int matchwidth, int matchheight, qboolean complain, qboolean mipmap)
-{
-	int texnum;
-	byte *data;
-	if (!(data = loadimagepixels (filename, complain, matchwidth, matchheight))) { 
-		Con_DPrintf("Cannot load image %s\n", filename);
-		return 0;
-	}
-	texnum = GL_LoadTexture (filename, image_width, image_height, data, mipmap, qtrue, 4);
-	free(data);
-	return texnum;
-}
-// Tomaz || TGA End
