+++ source/ctr/gl/gl_model.h
@@ -463,6 +463,5 @@ byte	*Mod_LeafPVS (mleaf_t *leaf, model_t *model);
 
 int GL_LoadTexture32 (char *identifier, int width, int height, byte *data, qboolean mipmap, qboolean alpha);		//Diabolickal HLBSP
 void BuildGammaTable (float g);
-int loadtextureimage (char* filename, int matchwidth, int matchheight, qboolean complain, qboolean mipmap);			//Diabolickal TGA
 
 #endif	// __MODEL__
