+++ source/nzportable_def.h
@@ -282,6 +282,8 @@ typedef struct
 
 #include RENDERER_FILE(main.h)
 
+#include "images.h"
+
 //=============================================================================
 
 // the host system specifies the base of the directory tree, the
