+++ source/psp/gu/gu_model.cpp
@@ -603,7 +603,7 @@ void Mod_LoadTextures (lump_t *l)
 		else
 		{
 			sprintf (texname, "textures/%s", mt->name);
-			tx->gl_texturenum = loadtextureimage (texname, 0, 0, qfalse, GU_LINEAR);
+			tx->gl_texturenum = loadtextureimage (texname, 0, 0, qfalse, GU_LINEAR, qfalse, qfalse);
 			if(tx->gl_texturenum == 0)
 			{
 				tx->gl_texturenum = GL_LoadTexture (mt->name, tx->width, tx->height, (byte *)(tx_pixels), qtrue, GU_LINEAR, level);
@@ -611,7 +611,7 @@ void Mod_LoadTextures (lump_t *l)
 /*
 		          //Crow_bar mult detail textures
 				  sprintf (detname, "gfx/detail/%s", mt->name);
-			      tx->dt_texturenum = loadtextureimage (detname, 0, 0, qfalse, GU_LINEAR);
+			      tx->dt_texturenum = loadtextureimage (detname, 0, 0, qfalse, GU_LINEAR, qfalse, qfalse);
 			      mapTextureNameList.push_back(tx->dt_texturenum);
 */
 			tx->dt_texturenum = 0;
@@ -2065,7 +2065,7 @@ void *Mod_LoadAllSkins (int numskins, daliasskintype_t *pskintype)
 			pheader->gl_texturenum[i][0] = 
 			pheader->gl_texturenum[i][1] = 
 			pheader->gl_texturenum[i][2] = 
-			pheader->gl_texturenum[i][3] = loadtextureimage("models/weapons/m1911/v_biatch.mdl_0", 0, 0, qtrue, GU_LINEAR);
+			pheader->gl_texturenum[i][3] = loadtextureimage("models/weapons/m1911/v_biatch.mdl_0", 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 
 			pskintype = (daliasskintype_t *)((byte *)(pskintype+1) + s);
 			return (void *)pskintype;
@@ -2078,7 +2078,7 @@ void *Mod_LoadAllSkins (int numskins, daliasskintype_t *pskintype)
 					pheader->gl_texturenum[i][0] =
 					pheader->gl_texturenum[i][1] =
 					pheader->gl_texturenum[i][2] =
-					pheader->gl_texturenum[i][3] = loadtextureimage("models/machines/v_perk.mdl_0", 0, 0, qtrue, GU_LINEAR);
+					pheader->gl_texturenum[i][3] = loadtextureimage("models/machines/v_perk.mdl_0", 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 					pskintype = (daliasskintype_t *)((byte *)(pskintype+1) + s);
 				} else if (!has_perk_revive && i == 0) {
 					pheader->gl_texturenum[i][0] =
@@ -2091,7 +2091,7 @@ void *Mod_LoadAllSkins (int numskins, daliasskintype_t *pskintype)
 					pheader->gl_texturenum[i][0] =
 					pheader->gl_texturenum[i][1] =
 					pheader->gl_texturenum[i][2] =
-					pheader->gl_texturenum[i][3] = loadtextureimage("models/machines/v_perk.mdl_1", 0, 0, qtrue, GU_LINEAR);
+					pheader->gl_texturenum[i][3] = loadtextureimage("models/machines/v_perk.mdl_1", 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 					pskintype = (daliasskintype_t *)((byte *)(pskintype+1) + s);
 				} else if (!has_perk_juggernog && i == 1) {
 					pheader->gl_texturenum[i][0] =
@@ -2104,7 +2104,7 @@ void *Mod_LoadAllSkins (int numskins, daliasskintype_t *pskintype)
 					pheader->gl_texturenum[i][0] =
 					pheader->gl_texturenum[i][1] =
 					pheader->gl_texturenum[i][2] =
-					pheader->gl_texturenum[i][3] = loadtextureimage("models/machines/v_perk.mdl_2", 0, 0, qtrue, GU_LINEAR);
+					pheader->gl_texturenum[i][3] = loadtextureimage("models/machines/v_perk.mdl_2", 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 					pskintype = (daliasskintype_t *)((byte *)(pskintype+1) + s);
 				} else if (!has_perk_speedcola && i == 2) {
 					pheader->gl_texturenum[i][0] =
@@ -2117,7 +2117,7 @@ void *Mod_LoadAllSkins (int numskins, daliasskintype_t *pskintype)
 					pheader->gl_texturenum[i][0] =
 					pheader->gl_texturenum[i][1] =
 					pheader->gl_texturenum[i][2] =
-					pheader->gl_texturenum[i][3] = loadtextureimage("models/machines/v_perk.mdl_3", 0, 0, qtrue, GU_LINEAR);
+					pheader->gl_texturenum[i][3] = loadtextureimage("models/machines/v_perk.mdl_3", 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 					pskintype = (daliasskintype_t *)((byte *)(pskintype+1) + s);
 				} else if (!has_perk_doubletap && i == 3) {
 					pheader->gl_texturenum[i][0] =
@@ -2130,7 +2130,7 @@ void *Mod_LoadAllSkins (int numskins, daliasskintype_t *pskintype)
 					pheader->gl_texturenum[i][0] =
 					pheader->gl_texturenum[i][1] =
 					pheader->gl_texturenum[i][2] =
-					pheader->gl_texturenum[i][3] = loadtextureimage("models/machines/v_perk.mdl_4", 0, 0, qtrue, GU_LINEAR);
+					pheader->gl_texturenum[i][3] = loadtextureimage("models/machines/v_perk.mdl_4", 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 					pskintype = (daliasskintype_t *)((byte *)(pskintype+1) + s);
 				} else if (!has_perk_staminup && i == 4) {
 					pheader->gl_texturenum[i][0] =
@@ -2143,7 +2143,7 @@ void *Mod_LoadAllSkins (int numskins, daliasskintype_t *pskintype)
 					pheader->gl_texturenum[i][0] =
 					pheader->gl_texturenum[i][1] =
 					pheader->gl_texturenum[i][2] =
-					pheader->gl_texturenum[i][3] = loadtextureimage("models/machines/v_perk.mdl_5", 0, 0, qtrue, GU_LINEAR);
+					pheader->gl_texturenum[i][3] = loadtextureimage("models/machines/v_perk.mdl_5", 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 					pskintype = (daliasskintype_t *)((byte *)(pskintype+1) + s);
 				} else if (!has_perk_flopper && i == 5) {
 					pheader->gl_texturenum[i][0] =
@@ -2156,7 +2156,7 @@ void *Mod_LoadAllSkins (int numskins, daliasskintype_t *pskintype)
 					pheader->gl_texturenum[i][0] =
 					pheader->gl_texturenum[i][1] =
 					pheader->gl_texturenum[i][2] =
-					pheader->gl_texturenum[i][3] = loadtextureimage("models/machines/v_perk.mdl_6", 0, 0, qtrue, GU_LINEAR);
+					pheader->gl_texturenum[i][3] = loadtextureimage("models/machines/v_perk.mdl_6", 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 					pskintype = (daliasskintype_t *)((byte *)(pskintype+1) + s);
 				} else if (!has_perk_deadshot && i == 6) {
 					pheader->gl_texturenum[i][0] =
@@ -2169,7 +2169,7 @@ void *Mod_LoadAllSkins (int numskins, daliasskintype_t *pskintype)
 					pheader->gl_texturenum[i][0] =
 					pheader->gl_texturenum[i][1] =
 					pheader->gl_texturenum[i][2] =
-					pheader->gl_texturenum[i][3] = loadtextureimage("models/machines/v_perk.mdl_7", 0, 0, qtrue, GU_LINEAR);
+					pheader->gl_texturenum[i][3] = loadtextureimage("models/machines/v_perk.mdl_7", 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 					pskintype = (daliasskintype_t *)((byte *)(pskintype+1) + s);
 				} else if (!has_perk_mulekick && i == 7) {
 					pheader->gl_texturenum[i][0] =
@@ -2198,7 +2198,7 @@ void *Mod_LoadAllSkins (int numskins, daliasskintype_t *pskintype)
 			pheader->gl_texturenum[i][0] = 
 			pheader->gl_texturenum[i][1] = 
 			pheader->gl_texturenum[i][2] = 
-			pheader->gl_texturenum[i][3] = loadtextureimage("models/weapons/v_papskin", 0, 0, qtrue, GU_LINEAR);
+			pheader->gl_texturenum[i][3] = loadtextureimage("models/weapons/v_papskin", 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 			pskintype = (daliasskintype_t *)((byte *)(pskintype+1) + s);
 			return (void *)pskintype;
 		}
@@ -2212,7 +2212,7 @@ void *Mod_LoadAllSkins (int numskins, daliasskintype_t *pskintype)
 			pheader->gl_texturenum[i][0] = 
 			pheader->gl_texturenum[i][1] = 
 			pheader->gl_texturenum[i][2] = 
-			pheader->gl_texturenum[i][3] = loadtextureimage (model2, 0, 0, qtrue, GU_LINEAR);
+			pheader->gl_texturenum[i][3] = loadtextureimage (model2, 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 
 			if (pheader->gl_texturenum[i][0] == 0)// did not find a matching TGA...
 			{
@@ -2241,7 +2241,7 @@ void *Mod_LoadAllSkins (int numskins, daliasskintype_t *pskintype)
 				COM_StripExtension(loadmodel->name, model);
 				snprintf(model2, 128, "%s_%i_%i", model, i, j);
 				pheader->gl_texturenum[i][j&3] =
-				loadtextureimage (model2, 0, 0, qfalse, GU_LINEAR);
+				loadtextureimage (model2, 0, 0, qfalse, GU_LINEAR, qfalse, qfalse);
 				
 				if (pheader->gl_texturenum[i][j&3] == 0)// did not find a matching TGA...
 				{
@@ -2630,8 +2630,6 @@ void Mod_LoadH2AliasModel (model_t *mod, void *buffer)
 Mod_LoadQ2AliasModel
 =================
 */
-int loadtextureimage (char* filename, int matchwidth, int matchheight, qboolean complain, int filter);
-
 void Mod_LoadQ2AliasModel (model_t *mod, void *buffer)
 {
 	//int                 numskins;
@@ -2701,7 +2699,7 @@ void Mod_LoadQ2AliasModel (model_t *mod, void *buffer)
 		pinskins = static_cast<char*>((void*)((int) pinmodel + LittleLong(pinmodel->ofs_skins)));
 		for (i = 0;i < pheader->num_skins;i++)
 		{
-			pheader->gl_texturenum[i] = loadtextureimage (pinskins, 0, 0, qtrue, GU_LINEAR);
+			pheader->gl_texturenum[i] = loadtextureimage (pinskins, 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 			pinskins += MD2MAX_SKINNAME;
 		}
 	}
@@ -2950,22 +2948,22 @@ void Mod_LoadQ3ModelTexture (char *identifier, int flags, int *gl_texnum)
 	char	loadpath[64];
 
 	Q_snprintfz (loadpath, sizeof(loadpath), "textures/q3models/%s", identifier);
-	*gl_texnum = loadtextureimage (loadpath, 0, 0, qtrue, GU_LINEAR);
+	*gl_texnum = loadtextureimage (loadpath, 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 
 	if (!*gl_texnum)
 	{
 		Q_snprintfz (loadpath, sizeof(loadpath), "textures/%s", identifier);
-		*gl_texnum = loadtextureimage (loadpath, 0, 0, qtrue, GU_LINEAR);
+		*gl_texnum = loadtextureimage (loadpath, 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 	}
 	if (!*gl_texnum)
 	{
 		Q_snprintfz (loadpath, sizeof(loadpath), "progs/%s", identifier);
-		*gl_texnum = loadtextureimage (loadpath, 0, 0, qtrue, GU_LINEAR);
+		*gl_texnum = loadtextureimage (loadpath, 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 	}
 	if (!*gl_texnum)
 	{
 		Q_snprintfz (loadpath, sizeof(loadpath), "maps/%s", identifier);
-		*gl_texnum = loadtextureimage (loadpath, 0, 0, qtrue, GU_LINEAR);
+		*gl_texnum = loadtextureimage (loadpath, 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 	}
 }
 
@@ -3504,7 +3502,7 @@ void * Mod_LoadSpriteFrame (void * pin, mspriteframe_t **ppframe, int framenum, 
 	{
 		COM_StripExtension(loadmodel->name, sprite);
 		snprintf (sprite2, 128, "%s.spr_%i", sprite, framenum);
-		pspriteframe->gl_texturenum = loadtextureimage (sprite2, 0, 0, qtrue, GU_LINEAR);
+		pspriteframe->gl_texturenum = loadtextureimage (sprite2, 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 		
 		if (pspriteframe->gl_texturenum == 0)// did not find a matching TGA...
 		{
@@ -3743,7 +3741,7 @@ qboolean Mod_LoadQ2SpriteModel (model_t *mod, void *buffer)
 
 		frame = psprite->frames[i].frameptr = static_cast<mspriteframe_t*>(Hunk_AllocName(sizeof(mspriteframe_t), loadname));
 
-		frame->gl_texturenum = loadtextureimage (pframetype->name, 0, 0, qtrue, GU_LINEAR);
+		frame->gl_texturenum = loadtextureimage (pframetype->name, 0, 0, qtrue, GU_LINEAR, qfalse, qfalse);
 
 		frame->width = LittleLong(pframetype->width);
 		frame->height = LittleLong(pframetype->height);
