+++ source/wii/gx/gx_main.h
@@ -423,7 +423,5 @@ void Sky_NewMap (void);
 void R_DrawSkyBox (void);
 void R_ClearSkyBox (void);
 void EmitBothSkyLayers (msurface_t *fa);
-byte* loadimagepixels (char* filename, int matchwidth, int matchheight, int reverseRGBA);
-int loadtextureimage (char* filename, int matchwidth, int matchheight, qboolean complain, qboolean keep);
 
 extern	qboolean	qmb_initialized;
