+++ source/wii/gx/gx_model.c
@@ -497,12 +497,12 @@ void Mod_LoadTextures (lump_t *l)
 						COM_StripExtension (loadmodel->name + 5, mapname);
 						//sprintf (filename, "textures/%s/%s", mapname, tx->name);
 						// don't keep map textures
-						data_ext = loadtextureimage (filename, 0, 0, true, false);
+						data_ext = loadtextureimage (filename, 0, 0, true, 0, false, false);
 						if (data_ext <= 0)
 						{
 							//Con_Printf ("mpath %s\n", filename);
 							sprintf (filename, "textures/%s", tx->name);
-							data_ext = loadtextureimage (filename, 0, 0, true, false);
+							data_ext = loadtextureimage (filename, 0, 0, true, 0, false, false);
 							if (data_ext <= 0) {
 								//Con_Printf ("mpath %s\n", filename);
 								//Con_Printf("no texture found, replacing with white (fixme) %s\n", tx->name);
@@ -1742,12 +1742,12 @@ void *Mod_LoadAllSkins (int numskins, daliasskintype_t *pskintype)
 		pheader->gl_texturenum[0][0] = 
 		pheader->gl_texturenum[0][1] = 
 		pheader->gl_texturenum[0][2] = 
-		pheader->gl_texturenum[0][3] = loadtextureimage("models/weapons/m1911/v_biatch.mdl_0", 0, 0, true, true);
+		pheader->gl_texturenum[0][3] = loadtextureimage("models/weapons/m1911/v_biatch.mdl_0", 0, 0, true, 0, true, false);
 		
 		pheader->gl_texturenum[1][0] = 
 		pheader->gl_texturenum[1][1] = 
 		pheader->gl_texturenum[1][2] = 
-		pheader->gl_texturenum[1][3] = loadtextureimage("models/weapons/m1911/v_biatch.mdl_0", 0, 0, true, true);
+		pheader->gl_texturenum[1][3] = loadtextureimage("models/weapons/m1911/v_biatch.mdl_0", 0, 0, true, 0, true, false);
 
 		pskintype = (daliasskintype_t *)((byte *)(pskintype+1) + s);
 		return (void *)pskintype;
@@ -1768,7 +1768,7 @@ void *Mod_LoadAllSkins (int numskins, daliasskintype_t *pskintype)
 			pheader->gl_texturenum[i][0] =
 			pheader->gl_texturenum[i][1] =
 			pheader->gl_texturenum[i][2] =
-			pheader->gl_texturenum[i][3] = loadtextureimage(model2, 0, 0, true, true);
+			pheader->gl_texturenum[i][3] = loadtextureimage(model2, 0, 0, true, 0, true, false);
 
 			if (pheader->gl_texturenum[i][0] == 0) // did not find a matching TGA...
 			{
@@ -2039,13 +2039,13 @@ void * Mod_LoadSpriteFrame (void * pin, mspriteframe_t **ppframe, int framenum)
 
 	COM_StripExtension(loadmodel->name, sprite);
 	sprintf(sprite2, "%s.spr_%i", sprite, framenum);
-	pspriteframe->gl_texturenum = loadtextureimage(sprite2, 0, 0, true, true);
+	pspriteframe->gl_texturenum = loadtextureimage(sprite2, 0, 0, true, 0, true, false);
 	
 	if (pspriteframe->gl_texturenum <= 0) // did not find a matching TGA...
 	{
 		//COM_StripExtension(loadmodel->name, sprite);
 		//sprintf(sprite2, "%s.spr", sprite);
-		//pspriteframe->gl_texturenum = loadtextureimage(sprite2, 0, 0, true, true);
+		//pspriteframe->gl_texturenum = loadtextureimage(sprite2, 0, 0, true, 0, true, false);
 		int len = strlen(sprite2);
 		texname = sprite2 + len - 20;
 		if (pspriteframe->gl_texturenum <= 0) {
