+++ source/wii/gx/gx_textures.c
@@ -868,297 +868,3 @@ void GL_ClearTextureCache(void)
 
 	GX_InvalidateTexAll();
 }
-/*
-//
-//
-//
-//
-//
-//
-*/
-
-
-/*
-============
-COM_FileBase
-============
-*/
-// create a unique identifier for each externally loaded texture based on filename
-void tex_filebase (char *in, char *out)
-{
-	char *s, *s2;
-	
-	s = in + strlen(in) - 1;
-	
-	for (s2 = s ; s2 != in && *s2 && *s2 != '/' ; s2--)
-	;
-	
-	if (s-s2 < 2)
-		strcpy (out,"bad input");
-	else
-	{
-		strncpy (out,s2+1, (s+1)-(s2));
-		out[s-s2] = 0;
-	}
-}
-
-
-int		image_width;
-int		image_height;
-/*
-=================================================================
-
-PCX LOADING
-
-=================================================================
-*/
-typedef struct
-{
-    char	manufacturer;
-    char	version;
-    char	encoding;
-    char	bits_per_pixel;
-    unsigned short	xmin,ymin,xmax,ymax;
-    unsigned short	hres,vres;
-    unsigned char	palette[48];
-    char	reserved;
-    char	color_planes;
-    unsigned short	bytes_per_line;
-    unsigned short	palette_type;
-    char	filler[58];
-    unsigned char	data;			// unbounded
-} pcx_t;
-/*
-============
-LoadPCX
-============
-*/
-byte* LoadPCX (char* filename, int matchwidth, int matchheight)
-{
-	pcx_t	*pcx;
-	byte	*palette;
-	byte	*out;
-	byte	*pix, *image_rgba;
-	int		x, y;
-	int		dataByte, runLength;
-	byte	*pcx_data;
-	byte	*pic;
-
-//
-// parse the PCX file
-//
-
-	// Load the raw data into memory, then store it
-    pcx_data = COM_LoadFile(filename, 5);
-
-    pcx = (pcx_t *)pcx_data;
-	
-	pcx->xmin = LittleShort(pcx->xmin);
-    pcx->ymin = LittleShort(pcx->ymin);
-    pcx->xmax = LittleShort(pcx->xmax);
-    pcx->ymax = LittleShort(pcx->ymax);
-    pcx->hres = LittleShort(pcx->hres);
-    pcx->vres = LittleShort(pcx->vres);
-    pcx->bytes_per_line = LittleShort(pcx->bytes_per_line);
-    pcx->palette_type = LittleShort(pcx->palette_type);
-	
-	if (pcx->manufacturer != 0x0a
-		|| pcx->version != 5
-		|| pcx->encoding != 1
-		|| pcx->bits_per_pixel != 8
-		|| pcx->xmax >= 640
-		|| pcx->ymax >= 480)
-	{
-		Con_Printf ("Bad pcx file %i\n");
-		return 0;
-	}
-	
-	image_rgba = &pcx->data;
-
-	out = malloc ((pcx->ymax+1)*(pcx->xmax+1)*4);	
-	pic = out;
-	pix = out;
-
-	if (matchwidth && (pcx->xmax+1) != matchwidth)
-		return 0;
-	if (matchheight && (pcx->ymax+1) != matchheight)
-		return 0;
-	
-	palette = malloc(768);
-	memcpy (palette, (byte *)pcx + com_filesize - 768, 768);
-	
-	for (y=0 ; y<=pcx->ymax ; y++)
-	{
-		pix = out + 4*y*(pcx->xmax+1);
-		
-		for (x=0 ; x<=pcx->xmax ; )
-		{
-			dataByte = *image_rgba++;
-
-			if((dataByte & 0xC0) == 0xC0)
-			{
-				runLength = dataByte & 0x3F;
-				dataByte = *image_rgba++;
-			}
-			else
-				runLength = 1;
-
-			while(runLength-- > 0) {
-				pix[0] = palette[dataByte*3+0];
-				pix[1] = palette[dataByte*3+1];
-				pix[2] = palette[dataByte*3+2];
-				pix[3] = 255;
-				pix += 4;
-				x++;
-			}
-		}
-	}
-	
-	image_width = pcx->xmax+1;
-	image_height = pcx->ymax+1;
-	
-	free (palette);
-	free (pcx_data);
-	return pic;
-}
-
-/*small function to read files with stb_image - single-file image loader library.
-** downloaded from: https://raw.githubusercontent.com/nothings/stb/master/stb_image.h
-** only use jpeg+png formats, because tbh there's not much need for the others.
-** */
-#define STB_IMAGE_IMPLEMENTATION
-#define STBI_FAILURE_USERMSG
-#define STBI_ONLY_JPEG
-#define STBI_ONLY_PNG
-#define STBI_ONLY_TGA
-#define STBI_ONLY_PIC
-#include "../../stb_image.h"
-byte* loadimagepixels (char* filename, int matchwidth, int matchheight, int reverseRGBA)
-{
-	int bpp;
-	int width, height;
-	
-	byte* rgba_data;
-	
-	// Load the raw data into memory, then store it
-    rgba_data = COM_LoadFile(filename, 5);
-
-	if (rgba_data == NULL) {
-		Con_Printf("NULL: %s\n", filename);
-		return 0;
-	}
-
-    byte *image = stbi_load_from_memory(rgba_data, com_filesize, &width, &height, &bpp, 4);
-	
-	if(image == NULL) {
-		Sys_Error(" stbi failure: %s\n", stbi_failure_reason());
-		return 0;
-	}
-	
-	//set image width/height for texture uploads
-	image_width = width;
-	image_height = height;
-
-	free(rgba_data);
-
-	return image;
-}
-
-int loadtextureimage (char* filename, int matchwidth, int matchheight, qboolean complain, qboolean keep)
-{
-	int	f;
-	int texnum;
-	char basename[128], name[256];
-	char texname[32];
-	byte *data;
-	byte *c;
-
-	if (complain == false)
-		COM_StripExtension(filename, basename); // strip the extension to allow TGA
-	else
-		strcpy(basename, filename);
-
-	c = (byte*)basename;
-	while (*c)
-	{
-		if (*c == '*')
-			*c = '+';
-		c++;
-	}
-
-	//Try PCX	
-	snprintf (name, 132, "%s.pcx", basename);
-	COM_FOpenFile (name, (FILE **)&f);
-	if (f > 0) {
-		COM_CloseFile (f);
-		tex_filebase (filename, texname);	
-		data = LoadPCX (name, matchwidth, matchheight);
-		if (data == 0) {
-			Con_Printf("PCX: can't load %s\n", name);	
-			return 0;
-		}
-		texnum = GL_LoadTexture (texname, image_width, image_height, data, false, true, true, 4);	
-		free(data);
-		return texnum;
-	}	
-	//Try TGA
-	snprintf (name, 132, "%s.tga", basename);
-	COM_FOpenFile (name,(FILE **) &f);
-	if (f > 0){
-		COM_CloseFile (f);
-		tex_filebase (filename, texname);	
-		data = loadimagepixels (name, matchwidth, matchheight, 4);	
-		if (data == 0) {
-			Con_Printf("TGA: can't load %s\n", name);	
-			return 0;
-		}
-		texnum = GL_LoadTexture (texname, image_width, image_height, data, false, true, keep, 4);	
-		free(data);
-		return texnum;
-	}
-	//Try PNG
-	snprintf (name, 132, "%s.png", basename);
-	COM_FOpenFile (name, (FILE **)&f);
-	if (f > 0){
-		COM_CloseFile (f);
-		tex_filebase (filename, texname);	
-		data = loadimagepixels (name, matchwidth, matchheight, 1);
-		if (data == 0) {
-			Con_Printf("PNG: can't load %s\n", name);	
-			return 0;
-		}
-		texnum = GL_LoadTexture (texname, image_width, image_height, data, false, true, keep, 4);
-		free(data);
-		return texnum;
-	}
-	//Try JPEG
-	snprintf (name, 133, "%s.jpeg", basename);
-	COM_FOpenFile (name, (FILE **)&f);
-	if (f > 0){
-		COM_CloseFile (f);
-		tex_filebase (filename, texname);	
-		data = loadimagepixels (name, matchwidth, matchheight, 1);
-		if (data == 0) {
-			Con_Printf("JPEG: can't load %s\n", name);	
-			return 0;
-		}
-		texnum = GL_LoadTexture (texname, image_width, image_height, data, false, true, keep, 4);
-		free(data);
-		return texnum;
-	}
-	snprintf (name, 132, "%s.jpg", basename);
-	COM_FOpenFile (name, (FILE **)&f);
-	if (f > 0){
-		COM_CloseFile (f);
-		tex_filebase (filename, texname);	
-		data = loadimagepixels (name, matchwidth, matchheight, 1);
-		if (data == 0) {
-			Con_Printf("JPG: can't load %s\n", name);	
-			return 0;
-		}
-		texnum = GL_LoadTexture (texname, image_width, image_height, data, false, true, keep, 4);
-		free(data);
-		return texnum;
-	}
-	return 0;
-}
